Adding Background Gradient and Navigation Bar
Created a black to dark purple color gradient that spans the height of the page. Color(s) can quickly be changed later
once the website is more complete. Added naviation bar that will direct viewer to different pages if they choose. Working
on keeping the home link in the top left and the other links to the top right.