Added Professional Experience Page - Edited Top Bar
Top bar had weird padding causing the bottom scroll wheel to show. Fix was to remove the padding and keep the top and bottom margins. Added the
professional experience page including information on what I was doing at each role in the company. May want to change this to a more bullet pointed
structure in the future but for the minute can keep as paragraphs. Companies added were Travel and Leisure and Cargill, with the most recent being
the top company.